Fix RoadInventory.yearOfConstruction type mismatch (Integer -> BigDecimal)
Caught by actually booting the app against the real gnn_road_survey
database on db-primary (Hibernate schema validation): the DDL declares
year_of_construction as NUMERIC(6,0), but the entity mapped it as
Integer, which Hibernate expects as int4. Verified end to end after
the fix: app boots clean, login/heartbeat/logout work against the live
DB, and the RSAC-vs-Nagar-Nigam city-scoping (Section 5) behaves
correctly — an RSAC login without a requested city is rejected, one
with a city is scoped to it, and a NAGAR_NIGAM user's requested city
override is silently ignored in favor of their fixed profile city.

+++ src/main/java/com/gnn/roadsurvey/entity/RoadInventory.java
@@ -60,8 +60,10 @@ public class RoadInventory {
     @Column(name = "construction_scheme")
     private String constructionScheme;
 
+    // NUMERIC(6,0) in the DDL, not int4 — matches Hibernate schema validation
+    // against the real column type (caught by an actual boot against db-primary).
     @Column(name = "year_of_construction")
-    private Integer yearOfConstruction;
+    private BigDecimal yearOfConstruction;
 
     @Column(name = "nearby_landmark")
     private String nearbyLandmark;
@@ -160,8 +162,8 @@ public class RoadInventory {
     public String getConstructionScheme() { return constructionScheme; }
     public void setConstructionScheme(String constructionScheme) { this.constructionScheme = constructionScheme; }
 
-    public Integer getYearOfConstruction() { return yearOfConstruction; }
-    public void setYearOfConstruction(Integer yearOfConstruction) { this.yearOfConstruction = yearOfConstruction; }
+    public BigDecimal getYearOfConstruction() { return yearOfConstruction; }
+    public void setYearOfConstruction(BigDecimal yearOfConstruction) { this.yearOfConstruction = yearOfConstruction; }
 
     public String getNearbyLandmark() { return nearbyLandmark; }
     public void setNearbyLandmark(String nearbyLandmark) { this.nearbyLandmark = nearbyLandmark; }
